# docs/api/demo-bff-v1.openapi.yaml
openapi: 3.1.0
info:
  title: SlateFS Consumer Demo BFF
  version: 1.0.0
  description: Same-origin, session-authenticated facade. Tenant identity and upstream credentials are never caller supplied.
  license: { name: Apache-2.0, identifier: Apache-2.0 }
servers:
  - url: /
security:
  - cookieAuth: []
paths:
...
  /api/v1/login:
    post:
      summary: Authenticate a demo account and rotate its session
      operationId: loginDemoSession
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, enum: [alice, bob] }
                password: { type: string }
      responses:
        "200":
          description: Rotated authenticated session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/logout:
    post:
      summary: Invalidate the current browser session
      operationId: logoutDemoSession
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "204": { description: Session invalidated }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/volumes/{volume}/versioning/branches:
    parameters: [{ $ref: "#/components/parameters/Volume" }]
    get:
      summary: List version branches
      operationId: listDemoVersionBranches
      responses:
        {
          "200": { description: Branch list },
          default: { $ref: "#/components/responses/Error" },
        }
    post:
      summary: Create a version branch
      operationId: createDemoVersionBranch
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/BranchRequest" } },
            },
        }
      responses:
        {
          "201": { description: Branch created },
          default: { $ref: "#/components/responses/Error" },
        }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: slatefs_demo_session }
  parameters:
    Csrf:
      {
        name: X-CSRF-Token,
        in: header,
        required: true,
        schema: { type: string },
      }
    Volume:
      {
        name: volume,
        in: path,
        required: true,
        schema: { type: string, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$" },
      }
...
  schemas:
    Session:
      type: object
      required: [authenticated, csrfToken, capabilities]
      properties:
        authenticated: { type: boolean }
        user:
          {
            type: object,
            required: [username, displayName],
            properties:
              { username: { type: string }, displayName: { type: string } },
          }
        csrfToken: { type: string }
        expiresAt: { type: string, format: date-time }
        capabilities:
          {
            type: object,
            required: [snapshots, versions, collaboration, repository],
            properties:
              {
                snapshots: { type: boolean },
                versions: { type: boolean },
                collaboration: { type: boolean },
                repository: { type: boolean },
              },
          }
...
    BranchRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string }
        commit: { type: string }
...
    Error:
      {
        type: object,
        required: [error],
        properties:
          {
            error:
              {
                type: object,
                required: [code, message, request_id, details],
                properties:
                  {
                    code: { type: string },
                    message: { type: string },
                    request_id: { type: string },
                    details: { type: object, additionalProperties: true },
                  },
              },
          },
      }
  responses:
    Error:
      {
        description: Sanitized error response,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }